requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
